+++ src/app/idea/[id]/_components/IdeaDetailMetaView.tsx
@@ -5,9 +5,6 @@ import {AttachmentList} from '@/app/idea/[id]/_components/AttachmentList';
 import {IdeaVersionDetail} from '@/schemas/idea';
 import {useParams, useRouter} from 'next/navigation';
 import {ROUTES} from '@/constants/routes';
-import {useState} from 'react';
-import {ModalWrapper} from '@/components/common/wrappers/ModalWrapper';
-import {GlobalAlertModal} from '@/components/common/modal/GlobalAlertModal';
 
 interface IdeaDetailMetaViewProps {
   version: IdeaVersionDetail;
@@ -20,8 +17,6 @@ export const IdeaDetailMetaView = ({
 }: IdeaDetailMetaViewProps) => {
   const router = useRouter();
   const params = useParams();
-
-  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState<boolean>(false);
 
   const imageAttachments = version.images.map((img) => ({
     name: img.fileName,
@@ -47,29 +42,18 @@ export const IdeaDetailMetaView = ({
     router.push(`${ROUTES.IDEA}/${params.id}/edit`);
   };
 
-  const handleDelete = () => {
-    setIsDeleteModalOpen(true);
-  };
-
-  const handleConfirmDelete = () => {
-    // TODO: 삭제 API 호출
-
-    setIsDeleteModalOpen(false);
-  };
-
   return (
     <section className='border-gray flex w-full flex-col gap-12 rounded-[8px] border p-6'>
       <div className='flex flex-col gap-6'>
         <div className='flex flex-row items-center justify-between'>
           <strong className='text-2xl font-bold'>
             <span>ver.{version.versionNumber} :</span>
-            <span className='ml-3'>{version.shortDescription}</span>
+            <span className='ml-3'>{version.title}</span>
           </strong>
           <div className='flex flex-row gap-4'>
             {isLatestVersion && (
               <GlobalSmallButton text='수정하기' onClick={handleEditMode} />
             )}
-            <GlobalSmallButton text='삭제하기' onClick={handleDelete} />
           </div>
         </div>
         <p>
@@ -96,20 +80,6 @@ export const IdeaDetailMetaView = ({
           <AttachmentList attachments={linkAttachments} />
         )}
       </div>
-
-      {isDeleteModalOpen && (
-        <ModalWrapper
-          isOpen={isDeleteModalOpen}
-          onClose={() => setIsDeleteModalOpen(false)}>
-          <GlobalAlertModal
-            withdrawText='계속 작성하기'
-            strongText='정말 삭제하시겠습니까?'
-            content='한 번 삭제하면 다시 되돌릴 수 없어요.'
-            onClose={handleConfirmDelete}
-            onContinue={() => setIsDeleteModalOpen(false)}
-          />
-        </ModalWrapper>
-      )}
     </section>
   );
 };
